+++ app/src/main/java/woowacourse/shopping/data/repository/DefaultProductRepository.kt
@@ -3,7 +3,7 @@ package woowacourse.shopping.data.repository
 import kotlinx.collections.immutable.ImmutableList
 import kotlinx.collections.immutable.toImmutableList
 import woowacourse.shopping.data.source.remote.ProductRemoteDataSource
-import woowacourse.shopping.data.source.remote.dto.product.toDomain
+import woowacourse.shopping.data.source.remote.dto.product.mapper.toDomain
 import woowacourse.shopping.domain.model.Product
 import woowacourse.shopping.domain.repository.ProductRepository
 import kotlin.collections.map
