+++ app/src/main/java/woowacourse/shopping/data/source/remote/dto/product/ProductContent.kt
@@ -2,9 +2,6 @@ package woowacourse.shopping.data.source.remote.dto.product
 
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
-import woowacourse.shopping.domain.model.Money
-import woowacourse.shopping.domain.model.Product
-import woowacourse.shopping.domain.model.ProductName
 
 @Serializable
 data class ProductContent(
@@ -19,18 +16,3 @@ data class ProductContent(
     @SerialName("price")
     val price: Int,
 )
-
-fun ProductContent.toDomain(): Product =
-    Product(
-        id = this.id,
-        name =
-            ProductName(
-                this.name,
-            ),
-        price =
-            Money(
-                this.price.toLong(),
-            ),
-        imageUrl = this.imageUrl,
-        category = this.category,
-    )
